feat: add tests

+++ src/OpenJpegDotNet/Delegates.cs
@@ -11,7 +11,7 @@ namespace OpenJpegDotNet
     /// <param name="bytes">The length of <parameref name="buffer"/>.</param>
     /// <param name="userData">The data pointer to have beee passed to <see cref="OpenJpeg.StreamSetUserData"/>.</param>
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate int StreamRead(IntPtr buffer, long bytes, IntPtr userData);
+    public delegate ulong StreamRead(IntPtr buffer, ulong bytes, IntPtr userData);
 
     /// <summary>
     /// Callback function prototype for write.
@@ -20,7 +20,7 @@ namespace OpenJpegDotNet
     /// <param name="bytes">The length of <parameref name="buffer"/>.</param>
     /// <param name="userData">The data pointer to have beee passed to <see cref="OpenJpeg.StreamSetUserData"/>.</param>
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate int StreamWrite(IntPtr buffer, long bytes, IntPtr userData);
+    public delegate ulong StreamWrite(IntPtr buffer, ulong bytes, IntPtr userData);
 
     /// <summary>
     /// Callback function prototype for skip.
@@ -28,7 +28,7 @@ namespace OpenJpegDotNet
     /// <param name="bytes">The position to skip.</param>
     /// <param name="userData">The data pointer to have beee passed to <see cref="OpenJpeg.StreamSetUserData"/>.</param>
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate int StreamSkip(long bytes, IntPtr userData);
+    public delegate long StreamSkip(ulong bytes, IntPtr userData);
 
     /// <summary>
     /// Callback function prototype for seek.
@@ -36,7 +36,7 @@ namespace OpenJpegDotNet
     /// <param name="bytes">The position to seek.</param>
     /// <param name="userData">The data pointer to have beee passed to <see cref="OpenJpeg.StreamSetUserData"/>.</param>
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate int StreamSeek(long bytes, IntPtr userData);
+    public delegate int StreamSeek(ulong bytes, IntPtr userData);
 
     /// <summary>
     /// Callback function to free user data.
